+++ tests/test_wordvalue.py
@@ -1,5 +1,5 @@
 import unittest
-from game.scrabble import ScrabbleGame
+from game.board import Board
 from game.cell import Cell
 from game.models import Tile
 
@@ -11,7 +11,7 @@ class TestCalculateWordValue(unittest.TestCase):
             Cell(letter=Tile('S', 2)),
             Cell(letter=Tile('A', 1))
         ]
-        value = ScrabbleGame.calculate_word_value(word)
+        value = Board.calculate_word_value(word)
         self.assertEqual(value, 5)
 
     def test_with_letter_multiplier(self):
@@ -25,7 +25,7 @@ class TestCalculateWordValue(unittest.TestCase):
             ),
             Cell(letter=Tile('A', 1)),
         ]
-        value = ScrabbleGame.calculate_word_value(word)
+        value = Board.calculate_word_value(word)
         self.assertEqual(value, 7)
 
     def test_with_word_multiplier(self):
@@ -39,7 +39,7 @@ class TestCalculateWordValue(unittest.TestCase):
             ),
             Cell(letter=Tile('A', 1)),
         ]
-        value = ScrabbleGame.calculate_word_value(word)
+        value = Board.calculate_word_value(word)
         self.assertEqual(value, 10)
 
     def test_with_letter_word_multiplier(self):
@@ -57,7 +57,7 @@ class TestCalculateWordValue(unittest.TestCase):
             ),
             Cell(letter=Tile('A', 1)),
         ]
-        value = ScrabbleGame.calculate_word_value(word)
+        value = Board.calculate_word_value(word)
         self.assertEqual(value, 14)
 
     def test_with_different_word(self):
@@ -80,7 +80,7 @@ class TestCalculateWordValue(unittest.TestCase):
             ),
             Cell(letter=Tile('A', 1))
         ]
-        value = ScrabbleGame.calculate_word_value(word)
+        value = Board.calculate_word_value(word)
         self.assertEqual(value, 38)
 
 if __name__ == '__main__':