+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - more tests for scrabble_game
 - word validation, parsing and putting [logic missing, init idea]
 - cli.py, test_cli.py
+- tests for test_board
+- board functionality [validate word inside board, is empty]
 ### Added [03/10/23]
 - tests for letters.py
 - tests for test_player.py, logic for letter checking in player.py